back/requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.7
django-cors-middleware==1.3.1
django-oauth-toolkit==1.2.0
django-rest-framework==0.1.0
djangorestframework==3.8.2
idna==2.7
oauthlib==2.1.0
pytz==2018.5
requests==2.19.1
urllib3==1.23
